frontend: Refactor Maximum Price Breakdown modal to its own file
- Make the structure more readable by separating it into components
- Fix some issues with a few fields
