+++ frontend/src/features/apartment/ApartmentMaxPricePage.tsx
@@ -1,8 +1,8 @@
 import React, {useEffect, useState} from "react";
 
 import {FetchBaseQueryError} from "@reduxjs/toolkit/query";
-import {Button, Dialog, Fieldset, IconCheck, Tabs} from "hds-react";
-import {useNavigate, useParams} from "react-router-dom";
+import {Button, Dialog, Fieldset, IconCheck} from "hds-react";
+import {useParams} from "react-router-dom";
 import {useImmer} from "use-immer";
 
 import {
@@ -10,284 +10,14 @@ import {
     useGetHousingCompanyDetailQuery,
     useSaveApartmentMaximumPriceMutation,
 } from "../../app/services";
-import {
-    FormInputField,
-    ImprovementsTable,
-    NavigateBackButton,
-    QueryStateHandler,
-    SaveButton,
-} from "../../common/components";
+import {FormInputField, ImprovementsTable, NavigateBackButton, QueryStateHandler} from "../../common/components";
 import {
     IApartmentDetails,
     IApartmentMaximumPrice,
     IApartmentMaximumPriceWritable,
     IHousingCompanyDetails,
 } from "../../common/models";
-import {formatMoney, hitasToast} from "../../common/utils";
-
-const CalculationRowPrice = ({label, calculation}) => {
-    return (
-        <div
-            className={`confirmation-modal__calculation-row${
-                calculation.maximum ? " confirmation-modal__calculation-row--maximum" : ""
-            }`}
-        >
-            <label>{label}</label>
-            <p>{formatMoney(calculation.maximum_price)}</p>
-        </div>
-    );
-};
-
-const MaximumPriceModalContent = ({
-    calculation,
-    apartment,
-    setIsModalVisible,
-}: {
-    calculation: IApartmentMaximumPrice;
-    apartment: IApartmentDetails;
-    setIsModalVisible;
-}) => {
-    const navigate = useNavigate();
-    const [confirmMaximumPrice, {data, error, isLoading}] = useSaveApartmentMaximumPriceMutation();
-    const initiallyActiveTab = () => {
-        switch (calculation.index) {
-            case "market_price_index":
-                return 0;
-            case "construction_price_index":
-                return 1;
-            case "surface_area_price_ceiling":
-                return 2;
-            default:
-                return 0;
-        }
-    };
-    const handleConfirmButton = () => {
-        confirmMaximumPrice({
-            data: {confirm: true},
-            id: calculation.id,
-            apartmentId: apartment.id,
-            housingCompanyId: apartment.links.housing_company.id,
-        });
-        setIsModalVisible(true);
-    };
-    useEffect(() => {
-        if (!isLoading && !error && data && data.confirmed_at) {
-            hitasToast("Enimmäishinta vahvistettu!");
-            navigate(`/housing-companies/${apartment.links.housing_company.id}/apartments/${apartment.id}`);
-        }
-    }, [apartment, data, error, isLoading, navigate]);
-    const BreakdownValue = ({label, value, unit = "€"}: {label: string; value?: number | string; unit?: string}) => (
-        <div className="confirmation-modal__breakdown-row">
-            <label>{label}</label>
-            <p>{`${(value && value.toString()) || "0"} ${unit}`}</p>
-        </div>
-    );
-
-    return (
-        <>
-            <Dialog.Content>
-                <Tabs initiallyActiveTab={initiallyActiveTab()}>
-                    <Tabs.TabList>
-                        <Tabs.Tab>
-                            <CalculationRowPrice
-                                label="Markkinahintaindeksi"
-                                calculation={calculation.calculations.market_price_index}
-                            />
-                        </Tabs.Tab>
-                        <Tabs.Tab>
-                            <CalculationRowPrice
-                                label="Rakennuskustannusindeksi"
-                                calculation={calculation.calculations.construction_price_index}
-                            />
-                        </Tabs.Tab>
-                        <Tabs.Tab>
-                            <CalculationRowPrice
-                                label="Rajaneliöhinta"
-                                calculation={calculation.calculations.surface_area_price_ceiling}
-                            />
-                        </Tabs.Tab>
-                    </Tabs.TabList>
-                    <Tabs.TabPanel className="confirmation-modal__breakdown">
-                        <BreakdownValue
-                            label="Hankinta-arvo"
-                            value={calculation.calculations.market_price_index.calculation_variables.acquisition_price}
-                        />
-                        <BreakdownValue
-                            label="+ Rakennusaikaiset lisä- ja muutostyöt"
-                            value={
-                                calculation.calculations.market_price_index.calculation_variables
-                                    .additional_work_during_construction
-                            }
-                        />
-                        <BreakdownValue
-                            label="= Perushinta"
-                            value={calculation.calculations.market_price_index.calculation_variables.basic_price}
-                        />
-                        <BreakdownValue
-                            label="+ Indeksin muutos"
-                            value={calculation.calculations.market_price_index.calculation_variables.index_adjustment}
-                        />
-                        <BreakdownValue
-                            label="+ Huoneistokohtaiset parannukset"
-                            value={
-                                calculation.calculations.market_price_index.calculation_variables
-                                    .housing_company_improvements?.summary.value
-                            }
-                        />
-                        <BreakdownValue
-                            label="+ Osuus yhtiön parannuksista"
-                            value={
-                                calculation.calculations.market_price_index.calculation_variables
-                                    .housing_company_improvements?.summary.value_for_apartment
-                            }
-                        />
-                        <BreakdownValue
-                            label="= Osakkeiden velaton hinta"
-                            value={calculation.calculations.market_price_index.calculation_variables.debt_free_price}
-                        />
-                        <BreakdownValue
-                            label="+ Huoneistokohtaiset vähennykset"
-                            value={calculation.calculations.market_price_index.calculation_variables.debt_free_price_m2}
-                        />
-                        <BreakdownValue
-                            label={`- Osuus yhtiön lainoista (${calculation.calculations.construction_price_index.calculation_variables.calculation_date})`}
-                            value={
-                                calculation.calculations.market_price_index.calculation_variables
-                                    .apartment_share_of_housing_company_loans
-                            }
-                        />
-                        <BreakdownValue
-                            label="= Enimmäismyyntihinta"
-                            value={calculation.calculations.market_price_index.maximum_price}
-                        />
-                        <BreakdownValue
-                            label="Velaton hinta euroa/m²"
-                            value={calculation.calculations.market_price_index.calculation_variables.debt_free_price_m2}
-                        />
-                    </Tabs.TabPanel>
-                    <Tabs.TabPanel className="confirmation-modal__breakdown">
-                        <BreakdownValue
-                            label="Hankinta-arvo"
-                            value={
-                                calculation.calculations.construction_price_index.calculation_variables
-                                    .acquisition_price
-                            }
-                        />
-                        <BreakdownValue
-                            label="+ Rakennusaikaiset lisä- ja muutostyöt"
-                            value={
-                                calculation.calculations.construction_price_index.calculation_variables
-                                    .additional_work_during_construction
-                            }
-                        />
-                        <BreakdownValue
-                            label="= Perushinta"
-                            value={calculation.calculations.construction_price_index.calculation_variables.basic_price}
-                        />
-                        <BreakdownValue
-                            label="+ Indeksin muutos"
-                            value={
-                                calculation.calculations.construction_price_index.calculation_variables.index_adjustment
-                            }
-                        />
-                        <BreakdownValue
-                            label="+ Huoneistokohtaiset parannukset"
-                            value={
-                                calculation.calculations.construction_price_index.calculation_variables
-                                    .housing_company_improvements?.summary.value
-                            }
-                        />
-                        <BreakdownValue
-                            label="+ Osuus yhtiön parannuksista"
-                            value={
-                                calculation.calculations.construction_price_index.calculation_variables
-                                    .housing_company_improvements?.summary.value_for_apartment
-                            }
-                        />
-                        <BreakdownValue
-                            label="= Osakkeiden velaton hinta"
-                            value={
-                                calculation.calculations.construction_price_index.calculation_variables.debt_free_price
-                            }
-                        />
-                        <BreakdownValue
-                            label="+ Huoneistokohtaiset vähennykset"
-                            value={
-                                calculation.calculations.construction_price_index.calculation_variables
-                                    .debt_free_price_m2
-                            }
-                        />
-                        <BreakdownValue
-                            label={`- Osuus yhtiön lainoista (${calculation.calculations.construction_price_index.calculation_variables.calculation_date})`}
-                            value={
-                                calculation.calculations.construction_price_index.calculation_variables
-                                    .apartment_share_of_housing_company_loans
-                            }
-                        />
-                        <BreakdownValue
-                            label="= Enimmäismyyntihinta"
-                            value={calculation.calculations.construction_price_index.maximum_price}
-                        />
-                        <BreakdownValue
-                            label="Velaton hinta euroa/m²"
-                            value={
-                                calculation.calculations.construction_price_index.calculation_variables
-                                    .debt_free_price_m2
-                            }
-                        />
-                    </Tabs.TabPanel>
-                    <Tabs.TabPanel className="confirmation-modal__breakdown">
-                        <BreakdownValue
-                            label="Asunnon pinta-ala"
-                            value={
-                                calculation.calculations.surface_area_price_ceiling.calculation_variables.surface_area
-                            }
-                            unit="m²"
-                        />
-                        <BreakdownValue
-                            label={`* Rajaneliöhinta (${calculation.calculations.surface_area_price_ceiling.calculation_variables.calculation_date})`}
-                            value={
-                                calculation.calculations.surface_area_price_ceiling.calculation_variables
-                                    .calculation_date_value
-                            }
-                        />
-                        <BreakdownValue
-                            label={`- Osuus yhtiön lainoista (${calculation.calculations.construction_price_index.calculation_variables.calculation_date})`}
-                            value={
-                                calculation.calculations.market_price_index.calculation_variables
-                                    .apartment_share_of_housing_company_loans
-                            }
-                        />
-                        <BreakdownValue
-                            label="= Enimmäishinta"
-                            value={calculation.calculations.surface_area_price_ceiling.maximum_price}
-                        />
-                    </Tabs.TabPanel>
-                </Tabs>
-                <div className="valid-until">
-                    <BreakdownValue
-                        label="Vahvistettavan laskelman voimassaoloaika"
-                        value={calculation.valid_until}
-                        unit=""
-                    />
-                </div>
-            </Dialog.Content>
-            <Dialog.ActionButtons className="align-content-right">
-                <Button
-                    onClick={() => setIsModalVisible(false)}
-                    variant="secondary"
-                    theme="black"
-                >
-                    Peruuta
-                </Button>
-                <SaveButton
-                    onClick={() => handleConfirmButton()}
-                    isLoading={isLoading}
-                />
-            </Dialog.ActionButtons>
-        </>
-    );
-};
+import MaximumPriceModalContent from "./components/ApartmentMaximumPriceBreakdownModal";
 
 const MaximumPriceModalError = ({error, setIsModalVisible}) => {
     const nonFieldError = ((error as FetchBaseQueryError)?.data as {message?: string})?.message || "";
